requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
anthropic==0.25.8
Pillow==10.2.0
python-decouple==3.8
mysqlclient==2.2.4
django-filter==24.2
drf-spectacular==0.27.1
